Move patched camera input to separate file and remove this file from test coverage.
