+++ packages/@dcl/inspector/jest.config.js
@@ -23,7 +23,7 @@ module.exports = {
     }
   },
   collectCoverageFrom: ["src/**/*.ts"],
-  coveragePathIgnorePatterns: ["node_modules/", "src/lib/babylon/decentraland/gizmo-patch.ts"],
+  coveragePathIgnorePatterns: ["node_modules/", "src/lib/babylon/decentraland/gizmo-patch.ts", "src/lib/babylon/decentraland/camera-input-patch.ts"],
   verbose: true,
   testMatch: ["**/*.spec.(ts)"],
   testEnvironment: "jsdom",
